move habitatctl to pkg
