+++ pkg/habitatctl/commands/stop.go
